+++ src/main/java/duke/Ui.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.util.Scanner;
 
 /**
@@ -74,4 +76,3 @@ public class Ui {
                 "____________________________________________________________");
     }
 }
-
